Changed project structure to cobra-style.
